drivers: thermal: step_wise: Handle the temperature stable trend.
Step wise algorithm ignores the temperature stable trend, which can
cause a case where step wise will not clear the mitigation, when the
temperature is lower than hysteresis and stays stable throughout.

In order to avoid this, consider temperature stable trend similar to
temperature dropping trend. This will lower the mitigation when the
temperature is less than hysteresis.

Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Iff70bdc403b90921c18d8dd0f3e4aae3aa721a0e